refs #26726 ALFView abstracted out view from presenter
