+++ qt/scientific_interfaces/Direct/ALFView.h
@@ -4,35 +4,38 @@
 //     NScD Oak Ridge National Laboratory, European Spallation Source
 //     & Institut Laue - Langevin
 // SPDX - License - Identifier: GPL - 3.0 +
-#ifndef MANTIDQT_CUSTOMINTERFACES_ALFVIEWPRESENTER_H_
-#define MANTIDQT_CUSTOMINTERFACES_ALFVIEWPRESENTER_H_
+#ifndef MANTIDQT_CUSTOMINTERFACES_ALFVIEW_H_
+#define MANTIDQT_CUSTOMINTERFACES_ALFVIEW_H_
 
 #include "DllConfig.h"
 #include "MantidQtWidgets/Common/UserSubWindow.h"
 #include "ALFView_view.h"
+#include "ALFView_presenter.h"
 #include "ALFView_model.h"
+
 
 namespace MantidQt {
 	namespace CustomInterfaces {
 /** ALCInterface : Custom interface for Avoided Level Crossing analysis
  */
-        class MANTIDQT_DIRECT_DLL ALFView_presenter : public QObject{
-          Q_OBJECT
+class MANTIDQT_DIRECT_DLL ALFView : public API::UserSubWindow {
+  Q_OBJECT
+
 
 public:
-  ALFView_presenter( ALFView_view *view);
-  ~ALFView_presenter(){};
-  void initLayout();
+  ALFView(QWidget *parent=nullptr);
+  ~ALFView(){};
+  static std::string name() { return "ALF View"; }
+  static QString categoryInfo() { return "Direct"; }
 
-private slots:
-  void loadRunNumber(); 
-  void loadBrowsedFile(const std::string);
+protected:
+  void initLayout() override;
 
 private:
-  void loadAndAnalysis(const std::string &run);
+
   ALFView_view *m_view;
   //ALFView_model *m_model;
-  int m_currentRun;
+  ALFView_presenter *m_presenter;
 };
 }// customInterfaces
 }// MantidQt
@@ -40,4 +43,4 @@ private:
 
 
 
-#endif /* MANTIDQT_CUSTOMINTERFACES_ALFVIEWPRESENTER_H_ */
+#endif /* MANTIDQT_CUSTOMINTERFACES_ALFVIEW_H_ */
