+++ qt/scientific_interfaces/Direct/CMakeLists.txt
@@ -1,9 +1,11 @@
 set(SRC_FILES
+    ALFView.cpp
     ALFView_view.cpp
     ALFView_presenter.cpp
     ALFView_model.cpp)
 
 set(MOC_FILES
+    ALFView.h
     ALFView_view.h
     ALFView_presenter.h
     ALFView_model.h)
@@ -82,4 +84,4 @@ mtd_add_qt_library(TARGET_NAME MantidScientificInterfacesDirect
                    LINUX_INSTALL_RPATH
                      "\$ORIGIN/../../${LIB_DIR};\$ORIGIN/../../plugins/qt5/") 
 # Testing target
-add_subdirectory(test)
+# add_subdirectory(test)
